Updates to empire 0.10.0 release in empire_ami
A lot of other stuff for getting the various new features to work.
Note: The AMI here isn't the final one, I'll update it once
https://github.com/remind101/empire_ami/pull/16
is pulled in and a final set of AMIs is available.

(Right now I've just updated the us-east-1 AMI with one of the AMIs
I was testing with)