Fix Elasticache template
The Elasticache template hasn't worked since at least 66d49693.

That commit replaced all references to Conditions with references to
Variables, except it missed the reference to HasInternalHostname.

It added a call to all() that never worked because its arguments
weren't passed as a list.

Finally, it changed the type of the Subnets variable from
EC2SubnetIdList to str. SubnetGroup.SubnetIds is a list:
http://docs.aws.amazon.com/AWSCloudFormation/latest/UserGuide/aws-properties-elasticache-subnetgroup.html#cfn-elasticache-subnetgroup-subnetids