efs: add blueprint
Add an Elastic File System blueprint.

It creates FileSystem instances and its associated MountTargets in a
specified list of Subnets, which makes things quite a bit easier than
creating them individually.

While it is uncommon, EFS allows specifying the IP addreses of its
internal instances manually, so we allow that too.

Since EFSs are commonly accessed from many different machines, a
convenience AllowedCIDRs variable is provided, that makes it easy to add
inbound rules to the newly-created security group for the FS.