Correctly test sqs.Queue parameters in sqs.py
validate_queue() has a fixed list of valid sqs.Queue parameters
which is not in sync with troposphere's actual parameters.

Notably, it lacks QueueName, which makes it fail when using
FifoQueue=True -- since FifoQueue requires a name set.

Instead of listing the parameters, use directly the keys from
sqs.Queue.props for checking.

Also add a test for sqs.py.